+++ conda/_vendor/auxlib/type_coercion.py
@@ -231,8 +231,8 @@ def typify(value, type_hint=None):
         elif not (type_hint - (STRING_TYPES_SET | {bool})):
             return boolify(value, return_string=True)
         elif not (type_hint - (STRING_TYPES_SET | {NoneType})):
-            value = typify_str_no_hint(text_type(value))
-            return None if value is None else text_type(value)
+            value = text_type(value)
+            return None if value.lower() == 'none' else value
         elif not (type_hint - {bool, int}):
             return typify_str_no_hint(text_type(value))
         else: