+++ conda/cli/common.py
@@ -11,7 +11,7 @@ import sys
 
 from .. import console
 from .._vendor.auxlib.entity import EntityEncoder
-from ..base.constants import ROOT_ENV_NAME
+from ..base.constants import ROOT_ENV_NAME, CONDA_TARBALL_EXTENSION
 from ..base.context import context, get_prefix as context_get_prefix
 from ..common.compat import iteritems
 from ..common.constants import NULL
@@ -88,7 +88,7 @@ class Packages(Completer):
 
     def _get_items(self):
         # TODO: Include .tar.bz2 files for local installs.
-        from conda.core.index import get_index
+        from ..core.index import get_index
         args = self.parsed_args
         call_dict = dict(channel_urls=args.channel or (),
                          use_cache=True,
@@ -106,7 +106,7 @@ class InstalledPackages(Completer):
 
     @memoize
     def _get_items(self):
-        from conda.core.linked_data import linked
+        from ..core.linked_data import linked
         packages = linked(context.prefix_w_legacy_search)
         return [dist.quad[0] for dist in packages]
 
@@ -412,6 +412,7 @@ def ensure_override_channels_requires_channel(args, dashc=True):
             raise CondaValueError('--override-channels requires --channel'
                                   'or --use-local')
 
+
 def confirm(args, message="Proceed", choices=('yes', 'no'), default='yes'):
     assert default in choices, default
     if args.dry_run:
@@ -476,7 +477,12 @@ def name_prefix(prefix):
 
 def arg2spec(arg, json=False, update=False):
     try:
-        spec = MatchSpec(spec_from_line(arg), normalize=True)
+        # spec_from_line can return None, especially for the case of a .tar.bz2 extension and
+        #   a space in the path
+        _arg = spec_from_line(arg)
+        if _arg is None and arg.endswith(CONDA_TARBALL_EXTENSION):
+            _arg = arg
+        spec = MatchSpec(_arg, normalize=True)
     except:
         raise CondaValueError('invalid package specification: %s' % arg)
 
@@ -572,7 +578,7 @@ def get_index_trap(*args, **kwargs):
     Retrieves the package index, but traps exceptions and reports them as
     JSON if necessary.
     """
-    from conda.core.index import get_index
+    from ..core.index import get_index
     kwargs.pop('json', None)
     return get_index(*args, **kwargs)
 
@@ -602,7 +608,7 @@ def stdout_json_success(success=True, **kwargs):
 
 
 def handle_envs_list(acc, output=True):
-    from conda import misc
+    from .. import misc
 
     if output:
         print("# conda environments:")